Update README.md
Fixes links for `outputFile()` and `outputJson()` in the table of content.
List `options` in a bullet point list. The current format is hard to read as it doesn't put different options on a newline.
Removed `copySync()` and added `Sync` description to `copy()` to be more concise with the rest of the doc.
Added a section to explicitly show the difference between `Sync` and `Async` idioms.