Fixing problem when copying file with no write permission
The 3rd parameter passed to openSync is setting the mode for the destination file, and if the source does not have write permissions, then the destination wouldn't have write permissions. It is best to use the default mode (which is 0666) and set the permissions after closing the file for write.