Test that remove() ignores glob characters.
After looking at rimraf, I began to fear that node-fs-extra, which
is based on rimraf, might also automatically glob things. I find
this confusing when the API is being used as an overlay of the
node built-in fs module which never expands globs. This test
explicitly verifies that the remove() function does not follow
glob characters.